*MAJOR REWORK* Short desc: Changes to project folder structure. all indicators are now classes (that inherit from an IndicatorBase class).
Long description: First, the project structure was changed to be more
streamlined and more in-line with what one commonly see in other projects.

Second, the IndicatorBase class was created as a base class for all indicators
and all indicators are now child classes of that class. This was done to allow
for greater functionality and to more easily implement strategies.
The indicators.py module has been deleted as result of this, as it has been
made redundant