Short description: Renamed and reworked the custom ticker class which inherits from yf.Ticker to better suit the new approach to indicators. renamed the .py file
Long description: Since indicators are now a class, the method that created a
DataFrame in the TickerWithIndicators (now TickerAdditionalPricepoints) was rendered obsolete.
It has now been removed.

The additional_price_points method has been refactored into the history method.
(The class extends its parent history() method and adds the pricepoints to
the DataFrame)