Short description: Fixed bugs in EMA and RSI indicators. changed the wording on the history method in TickerAdditionalPricepoints for clarity
Long description: the EMA class was applying its calculations on the wrong
DataFrame (self.base_df rather than self.df) resulting in a KeyError.
The rsi method in RelativeStrengthIndex was also modifying the wrong dataframe.
The history method in TickerAdditionalPricepoints had the arguments on its
declaration changed to *args and **kwargs (same functionality)