Short desc: Reworked the implementation of the strategy. added a testing example in the tests folder.
Long description: Removed the BaseStrategy class and replace it with a purely functional approach.
implementing the class was becoming too convoluted and wasn't quite working as expected.
The strategy_test file now includes a sketch of how the backtesting of the strategy should
go. Added a docstring that explains the process