Short Desc: Major rework. Strategies re-implemented as classes, with each indicator now having an associated strategy with it
Long desc: Re-implemented the classes as strategies, this time with a different
approach, creating a BaseStrategy class and then individual child classes for each
indicator. Created the macd_strategy class and implemented its backtesting method
(TODO: Determine if a Mixed signal method for signal detection is needed and if so, implement it. Implement the optimization method)

Additionally, the __str__ method of each indicator was re-implemented and now
returns the name of the indicators + its parameters (ie: RSI(14)).
The MACD indicator now also calculates the PPO (percentage price oscillator). Basically,
the difference between the fast and the slow ema expressed in terms of %