REWORK: Reworked the methods in the BaseStrategy class and the class that inherit from it, in order to avoid duplicate code and to streamline the implementation of future children of BaseStrategy.
-BaseStrategy now has the get_signal_column method. This method
takes a buy and a sell condition and returns a series using np.select.
This should allow to more easily combine strategies in the future
by allowing to easily retrieving the signal column only.
-get_performance method implemented in BaseStrategy. It performs the
calculations that were performed by the old backtest_strategy method
in the children classes. Since those calculations are the streamline
regardless of the strategy, it makes little sense to have to implement
it in each child class instead of having it implemented directly on the
parent class.
-BaseStrategy now has create_conditions and backtest_strategy as
abstract methods that child classes must implement.

macd_strategy and bollinger_strategy were modified as a result of this.